+++ src/sdc11073/xml_types/xml_structure.py
@@ -9,11 +9,12 @@ Container properties represent values in xml nodes.
 from __future__ import annotations
 
 import copy
+import operator
 import time
 from abc import ABC, abstractmethod
 from collections import OrderedDict
 from datetime import date, datetime
-from typing import TYPE_CHECKING, Any, Callable, Protocol
+from typing import TYPE_CHECKING, Any, Callable
 
 from lxml import etree as etree_
 
@@ -802,14 +803,53 @@ class NodeTextQNameProperty(_ElementBase):
             sub_node.text = value
 
 
+def _compare_extension(left: Any, right: Any) -> bool:
+    if not isinstance(left, etree_._Element) or not isinstance(right, etree_._Element):
+        return left == right  # default comparison
+
+    # xml comparison
+    if left.tag != right.tag:  # compare expanded names
+        return False
+    if dict(left.attrib) != dict(right.attrib):  # unclear how lxml _Attrib compares
+        return False
+
+    # ignore comments
+    left_children = [child for child in left if not isinstance(child, etree_._Comment)]
+    right_children = [child for child in right if not isinstance(child, etree_._Comment)]
+
+    if len(left_children) != len(right_children):  # compare children count
+        return False
+    if len(left_children) == 0 and len(right_children) == 0:
+        if left.text != right.text:  # mixed content is not allowed. only compare text if there are no children
+            return False
+
+    return all(map(_compare_extension, left_children, right_children))  # compare children but keep order
+
+
 class ExtensionLocalValue:
-    def __init__(self, value: Any):
+
+    compare_method: Callable[[Any, Any], bool] = _compare_extension
+    """may be overwritten by user if a custom comparison behaviour is required"""
+
+    def __init__(self, value: OrderedDict | None):
         self.value = value or OrderedDict()
 
-    def __eq__(self, other: ExtensionLocalValue) -> bool:
+    def __eq__(self, other: ExtensionLocalValue | None) -> bool:
         if other is None:
             return len(self.value) == 0
-        return self.value == other.value
+        if not isinstance(other, self.__class__):
+            return NotImplemented
+
+        if len(self.value) != len(other.value):
+            return False
+        for key, value in self.value.items():
+            try:
+                other_value = other.value[key]
+            except KeyError:
+                return False
+            if not ExtensionLocalValue.compare_method(value, other_value):
+                return False
+        return all(map(operator.eq, self.value, other.value))  # also check order
 
 
 class ExtensionNodeProperty(_ElementBase):