+++ tests/test_pmtypes.py
@@ -1,8 +1,10 @@
+import collections
 import unittest
+from unittest import mock
 
-from lxml.etree import QName, fromstring
+import lxml
 
-from sdc11073.xml_types import pm_types
+from sdc11073.xml_types import pm_types, xml_structure
 
 
 class TestPmTypes(unittest.TestCase):
@@ -46,13 +48,13 @@ class TestPmTypes(unittest.TestCase):
                 </pm:Type>
               </pm:AllowedValue>
 """
-        node = fromstring(text.format(''))
+        node = lxml.etree.fromstring(text.format(''))
         allowed_value1 = pm_types.AllowedValue.from_node(node)
         self.assertEqual(allowed_value1.Value, '')
-        generated_node = allowed_value1.as_etree_node(QName('foo', 'bar'), {})
+        generated_node = allowed_value1.as_etree_node(lxml.etree.QName('foo', 'bar'), {})
         self.assertEqual('', generated_node[0].text)
 
-        node = fromstring(text.format('foobar'))
+        node = lxml.etree.fromstring(text.format('foobar'))
         allowed_value2 = pm_types.AllowedValue.from_node(node)
         self.assertEqual(allowed_value2.Value, 'foobar')
 
@@ -63,7 +65,211 @@ class TestPmTypes(unittest.TestCase):
                       <pm:Arg xmlns:dd="dummy">dd:Something</pm:Arg>
                   </pm:Argument>
         """
-        node = fromstring(text.format(''))
+        node = lxml.etree.fromstring(text.format(''))
         arg = pm_types.ActivateOperationDescriptorArgument.from_node(node)
         self.assertEqual(arg.ArgName, pm_types.CodedValue("202890"))
-        self.assertEqual(arg.Arg, QName("dummy", "Something"))
+        self.assertEqual(arg.Arg, lxml.etree.QName("dummy", "Something"))
+
+
+class TestExtensions(unittest.TestCase):
+
+    def test_compare_extensions(self):
+        xml = b"""
+        <pm:Identification xmlns:pm="http://standards.ieee.org/downloads/11073/11073-10207-2017/participant"
+                           Root="urn:uuid:90beab82-f160-4e2f-b3b2-ed8cfcf5e205"
+                           Extension="123.234.424">
+            <ext:Extension xmlns:ext="http://standards.ieee.org/downloads/11073/11073-10207-2017/extension">
+                <foo someattr="somevalue"/>
+                    <foo_child childattr="somechild"/>
+                <bar anotherattr="differentvalue"/>
+            </ext:Extension>
+        </pm:Identification>
+        """
+        self.assertNotEqual(lxml.etree.fromstring(xml), lxml.etree.fromstring(xml))
+        inst1 = pm_types.InstanceIdentifier.from_node(lxml.etree.fromstring(xml))
+        inst2 = pm_types.InstanceIdentifier.from_node(lxml.etree.fromstring(xml))
+        self.assertEqual(inst1.ExtExtension, inst2.ExtExtension)
+        self.assertEqual(inst1, inst2)
+
+        another_xml = b"""
+                <pm:Identification xmlns:pm="http://standards.ieee.org/downloads/11073/11073-10207-2017/participant"
+                                   Root="urn:uuid:90beab82-f160-4e2f-b3b2-ed8cfcf5e205"
+                                   Extension="123.234.424">
+                    <ext:Extension xmlns:ext="http://standards.ieee.org/downloads/11073/11073-10207-2017/extension">
+                        <foo someattr="somevalue"/>
+                        <bar anotherattr="differentvalue2"/>
+                    </ext:Extension>
+                </pm:Identification>
+                """
+        inst2 = pm_types.InstanceIdentifier.from_node(lxml.etree.fromstring(another_xml))
+        self.assertNotEqual(inst1.ExtExtension, inst2.ExtExtension)
+        self.assertNotEqual(inst1, inst2)
+
+    def test_order(self):
+        xml1 = b"""
+        <pm:Identification xmlns:pm="http://standards.ieee.org/downloads/11073/11073-10207-2017/participant"
+                           Root="urn:uuid:90beab82-f160-4e2f-b3b2-ed8cfcf5e205"
+                           Extension="123.234.424">
+            <ext:Extension xmlns:ext="http://standards.ieee.org/downloads/11073/11073-10207-2017/extension">
+                <foo someattr="somevalue"/>
+                <bar anotherattr="differentvalue"/>
+            </ext:Extension>
+        </pm:Identification>
+        """
+        xml2 = b"""
+        <pm:Identification xmlns:pm="http://standards.ieee.org/downloads/11073/11073-10207-2017/participant"
+                           Root="urn:uuid:90beab82-f160-4e2f-b3b2-ed8cfcf5e205"
+                           Extension="123.234.424">
+            <ext:Extension xmlns:ext="http://standards.ieee.org/downloads/11073/11073-10207-2017/extension">
+                <bar anotherattr="differentvalue"/>
+                <foo someattr="somevalue"/>
+            </ext:Extension>
+        </pm:Identification>
+        """
+        inst1 = pm_types.InstanceIdentifier.from_node(lxml.etree.fromstring(xml1))
+        inst2 = pm_types.InstanceIdentifier.from_node(lxml.etree.fromstring(xml2))
+        self.assertNotEqual(inst1.ExtExtension, inst2.ExtExtension)
+        self.assertNotEqual(inst1, inst2)
+
+    def test_fails_with_qname(self):
+        xml1 = lxml.etree.fromstring(b"""<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<ext:Extension xmlns:ext="http://standards.ieee.org/downloads/11073/11073-10207-2017/extension"
+        xmlns:what="123.456.789">
+        <what:ItIsNotKnown>
+                <what:Unknown>what:lorem</what:Unknown>
+        </what:ItIsNotKnown>
+</ext:Extension>""")
+        xml2 = lxml.etree.fromstring(b"""<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<ext:Extension xmlns:ext="http://standards.ieee.org/downloads/11073/11073-10207-2017/extension"
+        xmlns:who="123.456.789">
+        <who:ItIsNotKnown>
+                <who:Unknown>who:lorem</who:Unknown>
+        </who:ItIsNotKnown>
+</ext:Extension>""")
+        self.assertNotEqual(lxml.etree.tostring(xml1), lxml.etree.tostring(xml2))
+        inst1 = xml_structure.ExtensionLocalValue(collections.OrderedDict([(xml1.tag, xml1)]))
+        inst2 = xml_structure.ExtensionLocalValue(collections.OrderedDict([(xml2.tag, xml2)]))
+        self.assertNotEqual(inst1, inst2)
+
+    def test_ignore_namespaces(self):
+        xml1 = lxml.etree.fromstring(b"""<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<ext:Extension xmlns:ext="http://standards.ieee.org/downloads/11073/11073-10207-2017/extension" 
+xmlns:pm="http://standards.ieee.org/downloads/11073/11073-10207-2017/participant">
+<what:ItIsNotKnown xmlns:what="123.456.789"><what:Unknown>What does this mean?</what:Unknown></what:ItIsNotKnown>
+</ext:Extension>""")
+        xml2 = lxml.etree.fromstring(b"""<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<ext:Extension xmlns:ext="http://standards.ieee.org/downloads/11073/11073-10207-2017/extension" xmlns:what="123.456.789">
+<what:ItIsNotKnown><what:Unknown>What does this mean?</what:Unknown></what:ItIsNotKnown>
+</ext:Extension>""")
+        self.assertNotEqual(lxml.etree.tostring(xml1), lxml.etree.tostring(xml2))
+        inst1 = xml_structure.ExtensionLocalValue(collections.OrderedDict([(xml1.tag, xml1)]))
+        inst2 = xml_structure.ExtensionLocalValue(collections.OrderedDict([(xml2.tag, xml2)]))
+        self.assertEqual(inst1, inst2)
+
+    def test_different_length(self):
+        inst1 = xml_structure.ExtensionLocalValue(collections.OrderedDict([(mock.MagicMock(), mock.MagicMock()),
+                                                                           (mock.MagicMock(), mock.MagicMock())]))
+        inst2 = xml_structure.ExtensionLocalValue(collections.OrderedDict([(mock.MagicMock(), mock.MagicMock())]))
+        self.assertNotEqual(inst1, inst2)
+
+    def test_different_keys(self):
+        inst1 = xml_structure.ExtensionLocalValue(collections.OrderedDict([(1, mock.MagicMock())]))
+        inst2 = xml_structure.ExtensionLocalValue(collections.OrderedDict([(2, mock.MagicMock())]))
+        self.assertNotEqual(inst1, inst2)
+
+    def test_compare_non_xml(self):
+        inst1 = xml_structure.ExtensionLocalValue(collections.OrderedDict([(1, '1')]))
+        inst2 = xml_structure.ExtensionLocalValue(collections.OrderedDict([(1, '1')]))
+        self.assertEqual(inst1, inst2)
+
+        inst3 = xml_structure.ExtensionLocalValue(collections.OrderedDict([(1, '2')]))
+        self.assertNotEqual(inst1, inst3)
+
+    def test_ignore_comments(self):
+        xml1 = lxml.etree.fromstring(b"""<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<ext:Extension xmlns:ext="http://standards.ieee.org/downloads/11073/11073-10207-2017/extension">
+<what:ItIsNotKnown xmlns:what="123.456.789"><what:Unknown>What does this mean?</what:Unknown></what:ItIsNotKnown>
+<!--This is an xml comment and should be ignored during comparison-->
+</ext:Extension>""")
+        xml2 = lxml.etree.fromstring(b"""<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<ext:Extension xmlns:ext="http://standards.ieee.org/downloads/11073/11073-10207-2017/extension">
+<what:ItIsNotKnown xmlns:what="123.456.789"><what:Unknown>What does this mean?</what:Unknown></what:ItIsNotKnown>
+</ext:Extension>""")
+        inst1 = xml_structure.ExtensionLocalValue(collections.OrderedDict([(xml1.tag, xml1)]))
+        inst2 = xml_structure.ExtensionLocalValue(collections.OrderedDict([(xml2.tag, xml2)]))
+        self.assertEqual(inst1, inst2)
+
+    def test_custom_compare_method(self):
+        inst1 = xml_structure.ExtensionLocalValue(collections.OrderedDict([(1, '1')]))
+        inst2 = xml_structure.ExtensionLocalValue(collections.OrderedDict([(1, '2')]))
+        self.assertNotEqual(inst1, inst2)
+
+        def _my_comparer(_, __):
+            return True
+
+        orig_method = xml_structure.ExtensionLocalValue.compare_method
+        xml_structure.ExtensionLocalValue.compare_method = _my_comparer
+        self.assertEqual(inst1, inst2)
+        xml_structure.ExtensionLocalValue.compare_method = orig_method
+
+    def test_cdata(self):
+        xml1 = b"""<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+                        <ext:Extension xmlns:ext="http://standards.ieee.org/downloads/11073/11073-10207-2017/extension">
+                        <what:ItIsNotKnown xmlns:what="123.456.789">
+                        <![CDATA[<some test data & stuff>]]>
+                        <what:Unknown>What does this mean?<![CDATA[Test this CDATA section]]></what:Unknown></what:ItIsNotKnown>
+                        </ext:Extension>"""
+        xml2 = b"""<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+                        <ext:Extension xmlns:ext="http://standards.ieee.org/downloads/11073/11073-10207-2017/extension">
+                        <who:ItIsNotKnown xmlns:who="123.456.789">
+                        <![CDATA[<some test data & stuff>]]>
+                        <who:Unknown>What does this mean?<![CDATA[Test this CDATA section]]></who:Unknown></who:ItIsNotKnown>
+                        </ext:Extension>"""
+        xml1 = lxml.etree.fromstring(xml1)
+        xml2 = lxml.etree.fromstring(xml2)
+
+        inst1 = xml_structure.ExtensionLocalValue(collections.OrderedDict([(xml1.tag, xml1)]))
+        inst2 = xml_structure.ExtensionLocalValue(collections.OrderedDict([(xml2.tag, xml2)]))
+        self.assertEqual(inst1, inst2)
+
+    def test_comparison_subelements(self):
+        xml1 = b"""
+            <ext:Extension xmlns:ext="http://standards.ieee.org/downloads/11073/11073-10207-2017/extension">
+                <foo someattr="somevalue">
+                    <foo_subelement subelement="value"/>
+                </foo>
+                <bar anotherattr="differentvalue"/>
+            </ext:Extension>
+        """
+        xml2 = b"""
+            <ext:Extension xmlns:ext="http://standards.ieee.org/downloads/11073/11073-10207-2017/extension">
+                <foo someattr="somevalue">
+                    <foo_subelement subelement="fiff_value"/>
+                </foo>
+                <bar anotherattr="differentvalue"/>
+            </ext:Extension>
+        """
+        xml1 = lxml.etree.fromstring(xml1)
+        xml2 = lxml.etree.fromstring(xml2)
+
+        inst1 = xml_structure.ExtensionLocalValue(collections.OrderedDict([(xml1.tag, xml1)]))
+        inst2 = xml_structure.ExtensionLocalValue(collections.OrderedDict([(xml2.tag, xml2)]))
+        self.assertNotEqual(inst1, inst2)
+
+    def test_mixed_content_is_ignored(self):
+        xml1 = lxml.etree.fromstring(b"""<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<ext:Extension xmlns:ext="http://standards.ieee.org/downloads/11073/11073-10207-2017/extension"
+        xmlns:what="123.456.789">
+        <what:ItIsNotKnown><what:Unknown>what:lorem</what:Unknown></what:ItIsNotKnown>
+</ext:Extension>""")
+        xml2 = lxml.etree.fromstring(b"""<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<ext:Extension xmlns:ext="http://standards.ieee.org/downloads/11073/11073-10207-2017/extension"
+        xmlns:what="123.456.789">
+        <what:ItIsNotKnown>
+        dsafasdf
+        <what:Unknown>what:lorem</what:Unknown>
+        </what:ItIsNotKnown>
+</ext:Extension>""")
+        inst1 = xml_structure.ExtensionLocalValue(collections.OrderedDict([(xml1.tag, xml1)]))
+        inst2 = xml_structure.ExtensionLocalValue(collections.OrderedDict([(xml2.tag, xml2)]))
+        self.assertEqual(inst1, inst2)